poor gcc for solaris seems to have problems selecting the right convert_endian routine, so i thought i help a bit by casting the input parameter
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@9814 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csengine/objects/thing.cpp
@@ -203,18 +203,18 @@ void csThing::GenerateCacheName ()
   mf.Write ((char*)&l, 4);
   if (GetName ())
     mf.Write (GetName (), strlen (GetName ()));
-
-  l = convert_endian (QInt ((b.MinX () * 1000)+.5));
+  
+  l = convert_endian ((long)QInt ((b.MinX () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian (QInt ((b.MinY () * 1000)+.5));
+  l = convert_endian ((long)QInt ((b.MinY () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian (QInt ((b.MinZ () * 1000)+.5));
+  l = convert_endian ((long)QInt ((b.MinZ () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian (QInt ((b.MaxX () * 1000)+.5));
+  l = convert_endian ((long)QInt ((b.MaxX () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian (QInt ((b.MaxY () * 1000)+.5));
+  l = convert_endian ((long)QInt ((b.MaxY () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian (QInt ((b.MaxZ () * 1000)+.5));
+  l = convert_endian ((long)QInt ((b.MaxZ () * 1000)+.5));
   mf.Write ((char*)&l, 4);
 
   csMD5::Digest digest = csMD5::Encode (mf.GetData (), mf.GetSize ());
